requirements.txt
requests~=2.32.3
tqdm~=4.66.5
rembg~=2.0.59
pandas
numpy
opencv-python
configparser
albumentations
Pillow
